ColumnReaderFactory to create ColumnReaders
We can switch between old and new readers in this factory class
based on the session options
