+++ presto-parquet/src/main/java/com/facebook/presto/parquet/reader/LongDecimalColumnReader.java
@@ -24,7 +24,7 @@ import java.math.BigInteger;
 public class LongDecimalColumnReader
         extends AbstractColumnReader
 {
-    LongDecimalColumnReader(RichColumnDescriptor descriptor)
+    public LongDecimalColumnReader(RichColumnDescriptor descriptor)
     {
         super(descriptor);
     }
